Add files via upload
Self-reflection
Design Thinking is about the 5 process in solving a problem; Empathy, Define, Ideate, Prototype, Test. In this assignment, we decided to prototype a fast food ordering system to understand in details regarding the Design Thinking. In conclusion, we arrive that Design Thinking is an ideal process grasp essential understanding on proper system in order to develop higher quality system.